Move IVariableConstracts and NonglobalVariable to separate files
